ml/requirements.txt
PyPDF2>=3.0.0
python-docx>=1.1.0
pymongo>=4.6.0
scikit-learn>=1.4.0
